requirements.txt
click==6.6
cycler==0.10.0
Cython==0.24
flake8==2.6.2
Flask==0.11.1
Flask-MySQL==1.3
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
matplotlib==1.5.1
mccabe==0.5.0
MySQL-python==1.2.5
mysqlclient==1.3.7
neat-python==0.8
numpy==1.11.1
pycodestyle==2.0.0
pydub==0.16.4
pyflakes==1.2.3
pyparsing==2.1.5
python-dateutil==2.5.3
python-speech-features==0.3
pytz==2016.6
scipy==0.17.1
six==1.10.0
stft==0.5.2
Werkzeug==0.11.10
